+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/service/FilesItemService.kt
@@ -32,7 +32,7 @@ class FilesItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
     }
 
     override fun createPreClipItem(
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         identifier: String,
         transferable: Transferable,
@@ -47,7 +47,7 @@ class FilesItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
 
     override fun doLoadRepresentation(
         transferData: Any,
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         dataFlavor: DataFlavor,
         dataFlavorMap: Map<String, List<DataFlavor>>,
